Moved resource types to their own file
