+++ src/lib.rs
@@ -1,50 +1,15 @@
+mod resource;
+pub use resource::*;
+
 use std::{collections::HashMap, any::TypeId};
 
 use atomic_refcell::*;
-use downcast_rs::{Downcast, impl_downcast};
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub enum AccessError {
     NoSuchResource,
     AlreadyBorrowed,
 }
-
-pub trait Resource: Downcast + 'static {}
-
-impl<T> Resource for T where T: 'static {}
-
-impl_downcast!(Resource);
-
-struct ResourceCell {
-    inner: AtomicRefCell<Box<dyn Resource>>,
-}
-
-impl ResourceCell {
-    fn new(resource: Box<dyn Resource>) -> Self {
-        Self {
-            inner: AtomicRefCell::new(resource),
-        }
-    }
-
-    fn into_inner(self) -> Box<dyn Resource> {
-        self.inner.into_inner()
-    }
-
-    pub fn try_borrow<T: Resource>(&self) -> Result<AtomicRef<T>, AccessError> {
-        match self.inner.try_borrow() {
-            Ok(borrow) => Ok(AtomicRef::map(borrow, |inner| inner.downcast_ref().unwrap())),
-            Err(_) => Err(AccessError::AlreadyBorrowed),
-        }
-    }
-
-    pub fn try_borrow_mut<T: Resource>(&self) -> Result<AtomicRefMut<T>, AccessError> {
-        match self.inner.try_borrow_mut() {
-            Ok(borrow) => Ok(AtomicRefMut::map(borrow, |inner| inner.downcast_mut().unwrap())),
-            Err(_) => Err(AccessError::AlreadyBorrowed),
-        }
-    }
-}
-
 
 /// Provides a shared [`Resource`] container.
 #[derive(Default)]
